# Optimize volume integral kernels for shock capturing (frequent use)