--- a/profile/profiling_kernel.jl
+++ b/profile/profiling_kernel.jl
@@ -4,23 +4,24 @@ using Trixi, TrixiCUDA
 RealT = Float32
 
 # Set up the problem
-advection_velocity = 1.0f0
-equations = LinearScalarAdvectionEquation1D(advection_velocity)
+equations = CompressibleEulerEquations1D(1.4f0)
 
-solver = DGSEM(polydeg = 3, surface_flux = flux_lax_friedrichs, RealT = RealT)
+initial_condition = initial_condition_weak_blast_wave
 
-coordinates_min = -1.0f0
-coordinates_max = 1.0f0
+volume_flux = flux_ranocha
+solver_gpu = DGSEMGPU(polydeg = 3, surface_flux = flux_ranocha,
+                      volume_integral = VolumeIntegralFluxDifferencing(volume_flux),
+                      RealT = RealT)
 
+coordinates_min = (-2.0f0,)
+coordinates_max = (2.0f0,)
 mesh = TreeMesh(coordinates_min, coordinates_max,
-                initial_refinement_level = 4,
-                n_cells_max = 30_000, RealT = RealT)
+                initial_refinement_level = 5,
+                n_cells_max = 10_000, RealT = RealT)
 
-# Cache initialization
-semi_gpu = SemidiscretizationHyperbolicGPU(mesh, equations, initial_condition_convergence_test,
-                                           solver)
+semi_gpu = SemidiscretizationHyperbolicGPU(mesh, equations, initial_condition, solver_gpu)
 
-tspan_gpu = (0.0f0, 1.0f0)
+tspan_gpu = (0.0f0, 0.4f0)
 t_gpu = 0.0f0
 
 # Semi on GPU
@@ -31,8 +32,10 @@ source_terms_gpu = semi_gpu.source_terms
 
 # ODE on GPU
 ode_gpu = semidiscretizeGPU(semi_gpu, tspan_gpu)
-u_gpu = copy(ode_gpu.u0)
-du_gpu = similar(u_gpu)
+u_gpu_ = copy(ode_gpu.u0)
+du_gpu_ = similar(u_gpu_)
+u_gpu = TrixiCUDA.wrap_array(u_gpu_, mesh_gpu, equations_gpu, solver_gpu, cache_gpu)
+du_gpu = TrixiCUDA.wrap_array(du_gpu_, mesh_gpu, equations_gpu, solver_gpu, cache_gpu)
 
 # Reset du and volume integral
 TrixiCUDA.cuda_volume_integral!(du_gpu, u_gpu, mesh_gpu,