--- a/src/solvers/dg_1d.jl
+++ b/src/solvers/dg_1d.jl
@@ -404,6 +404,79 @@ function volume_flux_integral_dgfv_kernel!(du, u, alpha, atol, derivative_split,
     # Allocate dynamic shared memory
     shmem_split = @cuDynamicSharedMem(eltype(du), (tile_width, tile_width))
     offset += sizeof(eltype(du)) * tile_width^2
+    shmem_fstar = @cuDynamicSharedMem(eltype(du), (size(du, 1), tile_width + 1),
+                                      offset)
+    offset += sizeof(eltype(du)) * size(du, 1) * (tile_width + 1)
+    shmem_value = @cuDynamicSharedMem(eltype(du), (size(du, 1), tile_width),
+                                      offset)
+
+    # Get thread and block indices only we need to save registers
+    ty = threadIdx().y
+    k = (blockIdx().z - 1) * blockDim().z + threadIdx().z
+
+    # Load global `derivative_split` into shared memory
+    for ty2 in axes(du, 2)
+        # Transposed access
+        @inbounds shmem_split[ty2, ty] = derivative_split[ty, ty2]
+    end
+
+    # Get variables for computation
+    @inbounds alpha_element = alpha[k]
+    dg_only = isapprox(alpha_element, 0, atol = atol)
+
+    # Compute FV volume fluxes
+    u_node = get_node_vars(u, equations, ty, k)
+    if ty + 1 <= tile_width
+        flux_fv_node = volume_flux_fv(u_node,
+                                      get_node_vars(u, equations, ty + 1, k),
+                                      1, equations)
+    end
+
+    # Initialize the values
+    for tx in axes(du, 1)
+        @inbounds begin
+            # Initialize `du` with zeros
+            shmem_value[tx, ty] = zero(eltype(du))
+            # Initialize `fstar` side columes with zeros 
+            shmem_fstar[tx, 1] = zero(eltype(du))
+            shmem_fstar[tx, tile_width + 1] = zero(eltype(du))
+        end
+
+        if ty + 1 <= tile_width
+            # Set with FV volume fluxes
+            @inbounds shmem_fstar[tx, ty + 1] = flux_fv_node[tx] * (1 - dg_only)
+        end
+    end
+
+    sync_threads()
+
+    # Contribute FV to the volume integrals
+    for tx in axes(du, 1)
+        @inbounds shmem_value[tx, ty] += alpha_element * inverse_weights[ty] *
+                                         (shmem_fstar[tx, ty + 1] - shmem_fstar[tx, ty]) * (1 - dg_only)
+    end
+
+    # Compute DG volume fluxes
+    for thread in 1:tile_width
+        volume_flux_node = volume_flux_dg(u_node,
+                                          get_node_vars(u, equations, thread, k),
+                                          1, equations)
+
+        # Contribute DG to the volume integrals
+        for tx in axes(du, 1)
+            @inbounds shmem_value[tx, ty] += derivative_split[ty, thread] *
+                                             (1 - isequal(ty, thread)) * # set diagonal elements to zeros
+                                             volume_flux_node[tx] * dg_only +
+                                             (1 - alpha_element) * derivative_split[ty, thread] *
+                                             (1 - isequal(ty, thread)) * # set diagonal elements to zeros
+                                             volume_flux_node[tx] * (1 - dg_only)
+        end
+    end
+
+    # Finalize the values
+    for tx in axes(du, 1)
+        @inbounds du[tx, ty, k] = shmem_value[tx, ty]
+    end
 
     return nothing
 end
@@ -896,36 +969,51 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{1}, nonconservative_terms::
     derivative_split = dg.basis.derivative_split
     inverse_weights = dg.basis.inverse_weights
 
-    fstar1_L = cache.fstar1_L
-    fstar1_R = cache.fstar1_R
-
     # TODO: Get copies of `u` and `du` on both device and host
     # FIXME: Scalar indexing on GPU arrays caused by using GPU cache
     alpha = indicator(Array(u), mesh, equations, dg, cache) # GPU cache
     alpha = CuArray(alpha)
     atol = max(100 * eps(RealT), eps(RealT)^convert(RealT, 0.75f0))
 
-    volume_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
+    thread_per_block = size(du, 2)
+    if thread_per_block > MAX_THREADS_PER_BLOCK
+        # TODO: Remove `fstar` from cache initialization
+        fstar1_L = cache.fstar1_L
+        fstar1_R = cache.fstar1_R
 
-    volume_flux_dgfv_kernel = @cuda launch=false volume_flux_dgfv_kernel!(volume_flux_arr, fstar1_L,
-                                                                          fstar1_R, u,
-                                                                          alpha, atol,
-                                                                          equations, volume_flux_dg,
-                                                                          volume_flux_fv)
-    volume_flux_dgfv_kernel(volume_flux_arr, fstar1_L, fstar1_R, u, alpha, atol, equations,
-                            volume_flux_dg, volume_flux_fv;
-                            kernel_configurator_2d(volume_flux_dgfv_kernel, size(u, 2)^2,
-                                                   size(u, 3))...)
+        volume_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
 
-    volume_integral_dgfv_kernel = @cuda launch=false volume_integral_dgfv_kernel!(du, alpha,
-                                                                                  derivative_split,
-                                                                                  inverse_weights,
-                                                                                  volume_flux_arr,
-                                                                                  fstar1_L, fstar1_R,
-                                                                                  atol, equations)
-    volume_integral_dgfv_kernel(du, alpha, derivative_split, inverse_weights, volume_flux_arr,
-                                fstar1_L, fstar1_R, atol, equations;
-                                kernel_configurator_3d(volume_integral_dgfv_kernel, size(du)...)...)
+        volume_flux_dgfv_kernel = @cuda launch=false volume_flux_dgfv_kernel!(volume_flux_arr, fstar1_L,
+                                                                              fstar1_R, u,
+                                                                              alpha, atol,
+                                                                              equations, volume_flux_dg,
+                                                                              volume_flux_fv)
+        volume_flux_dgfv_kernel(volume_flux_arr, fstar1_L, fstar1_R, u, alpha, atol, equations,
+                                volume_flux_dg, volume_flux_fv;
+                                kernel_configurator_2d(volume_flux_dgfv_kernel, size(u, 2)^2,
+                                                       size(u, 3))...)
+
+        volume_integral_dgfv_kernel = @cuda launch=false volume_integral_dgfv_kernel!(du, alpha,
+                                                                                      derivative_split,
+                                                                                      inverse_weights,
+                                                                                      volume_flux_arr,
+                                                                                      fstar1_L, fstar1_R,
+                                                                                      atol, equations)
+        volume_integral_dgfv_kernel(du, alpha, derivative_split, inverse_weights, volume_flux_arr,
+                                    fstar1_L, fstar1_R, atol, equations;
+                                    kernel_configurator_3d(volume_integral_dgfv_kernel, size(du)...)...)
+    else
+        shmem_size = (size(du, 2)^2 + size(du, 1) * (size(du, 2) + 1) +
+                      size(du, 1) * size(du, 2)) * sizeof(RealT)
+        threads = (1, size(du, 2), 1)
+        blocks = (1, 1, size(du, 3))
+        @cuda threads=threads blocks=blocks shmem=shmem_size volume_flux_integral_dgfv_kernel!(du, u, alpha, atol,
+                                                                                               derivative_split,
+                                                                                               inverse_weights,
+                                                                                               equations,
+                                                                                               volume_flux_dg,
+                                                                                               volume_flux_fv)
+    end
 
     return nothing
 end
@@ -941,14 +1029,14 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{1}, nonconservative_terms::
     derivative_split = dg.basis.derivative_split
     inverse_weights = dg.basis.inverse_weights
 
-    fstar1_L = cache.fstar1_L
-    fstar1_R = cache.fstar1_R
-
     # TODO: Get copies of `u` and `du` on both device and host
     # FIXME: Scalar indexing on GPU arrays caused by using GPU cache
     alpha = indicator(Array(u), mesh, equations, dg, cache) # GPU cache
     alpha = CuArray(alpha)
     atol = max(100 * eps(RealT), eps(RealT)^convert(RealT, 0.75f0))
+
+    fstar1_L = cache.fstar1_L
+    fstar1_R = cache.fstar1_R
 
     volume_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
     noncons_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))