Continue #120 but focus on the frequent use cases without conservative terms.

These cases require high-level optimization to achieve high performance. Benchmark results should be provided for both before and after optimization prior to merging.

Checklist:
- [x] 1D without conservative terms
- [x] 2D without conservative terms
- [x] 3D without conservative terms