@huiyuxie 2025-01-23
Benchmarks are done based on Euler equations with shock capturing.