@huiyuxie 2025-01-23
1D does not earn much performance gain.

Before the optimization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 525 samples with 1 evaluation.
 Range (min … max):  7.543 ms … 19.418 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     8.884 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.515 ms ±  1.884 ms  ┊ GC (mean ± σ):  0.15% ± 1.47%

      ▆▇█▆▄
  ▃▄▆▇██████▅▅▄▆▄▃▄▄▄▄▃▃▄▄▃▃▂▃▂▂▁▂▃▂▂▂▃▁▁▂▂▂▂▂▁▁▂▁▂▁▁▂▁▂▂▂▂▂ ▃
  7.54 ms        Histogram: frequency by time        17.7 ms <

 Memory estimate: 421.58 KiB, allocs estimate: 3923.
```
After the optmization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 530 samples with 1 evaluation.
 Range (min … max):  7.493 ms … 20.489 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     8.861 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.449 ms ±  1.765 ms  ┊ GC (mean ± σ):  0.14% ± 1.55%

      ▁▃█▅▄▂
  ▂▂▃▅██████▆▅▄▃▃▃▃▃▃▃▄▃▄▃▃▃▃▂▂▃▃▂▁▂▁▁▁▂▁▂▂▂▁▁▂▂▃▁▂▁▂▁▁▂▁▁▁▂ ▃
  7.49 ms        Histogram: frequency by time        17.3 ms <

 Memory estimate: 418.94 KiB, allocs estimate: 3880.
```