@huiyuxie 2025-01-24
By investigation, the bad performance is caused by https://github.com/trixi-gpu/TrixiCUDA.jl/issues/59.