@huiyuxie 2025-01-24
The indicator function impacts the benchmark results for the kernel part (seconds dominating over microseconds), but the actual timing can be obtained through Nsight Compute, which we will provide later.