@huiyuxie 2025-01-24
Profiling for 1D (time only)
![Screenshot 2025-01-24 104642](https://github.com/user-attachments/assets/cad14a57-8323-4179-af6e-0d78fcaf65e3)

Profiling for 2D (time only)
![Screenshot 2025-01-24 101756](https://github.com/user-attachments/assets/05188d41-21d4-4edf-b597-98f509e7be0a)

Profiling for 3D (time only)
![Screenshot 2025-01-24 104343](https://github.com/user-attachments/assets/dac0c768-e836-4523-b800-beaff8ae08c7)